// apps/mobile/src/constants/theme.d.ts
export declare const colors: {
    light: {
        primary: string;
        primaryDark: string;
        secondary: string;
        background: string;
        surface: string;
        surfaceVariant: string;
        text: string;
        textSecondary: string;
        success: string;
        error: string;
        border: string;
    };
    dark: {
        primary: string;
        primaryDark: string;
        secondary: string;
        background: string;
        surface: string;
        surfaceVariant: string;
        text: string;
        textSecondary: string;
        success: string;
        error: string;
        border: string;
    };
};
export declare const spacing: {
    xs: number;
    sm: number;
    md: number;
    lg: number;
    xl: number;
};
export declare const borderRadius: {
    sm: number;
    md: number;
    lg: number;
    xl: number;
    full: number;
};
export declare const fontSize: {
    xs: number;
    sm: number;
    md: number;
    lg: number;
    xl: number;
    '2xl': number;
    '3xl': number;
};
//# sourceMappingURL=theme.d.ts.map